python backend/requirements.txt
django
futures
lxml
mysqlclient
pip
setuptools
requests
youtube-dl
gunicorn
pydub
djangorestframework
markdown
django-filter